+++ brownie/datastructures.py
@@ -33,8 +33,18 @@ del Missing
 
 def iter_multi_items(mapping):
     """
-    Iterates over the items of the given `mapping` yielding keys and values
-    without dropping any from more complex datastructures.
+    Iterates over the items of the given `mapping`.
+
+    If a key has multiple values a ``(key, value)`` item is yielded for each::
+
+        >>> for key, value in iter_multi_items({1: [2, 3]}):
+        ...     print key, value
+        1 2
+        1 3
+        >>> for key, value in iter_multi_items(MultiDict({1: [2, 3]})):
+        ...     print key, value
+        1 2
+        1 3
     """
     if isinstance(mapping, MultiDict):
         for item in mapping.iteritems(multi=False):
@@ -64,13 +74,83 @@ class ImmutableDictMixin(object):
     __setitem__ = __delitem__ = setdefault = update = pop = popitem = clear = \
         raise_immutable
 
+    def __repr__(self):
+        content = dict.__repr__(self) if self else ''
+        return '%s(%s)' % (self.__class__.__name__, content)
+
 
 class ImmutableDict(ImmutableDictMixin, dict):
     """An immutable :class:`dict`."""
 
+
+class CombinedDictMixin(object):
+    @classmethod
+    def fromkeys(cls, keys, value=None):
+        raise TypeError('cannot create %r instances with .fromkeys()' %
+            cls.__class__.__name__
+        )
+
+    def __init__(self, dicts=None):
+        #: The list of combined dictionaries.
+        self.dicts = [] if dicts is None else list(dicts)
+
+    def __getitem__(self, key):
+        for d in self.dicts:
+            if key in d:
+                return d[key]
+        raise KeyError(key)
+
+    def get(self, key, default=None):
+        try:
+            return self[key]
+        except KeyError:
+            return default
+
+    def __iter__(self):
+        return unique(chain.from_iterable(d.iterkeys() for d in self.dicts))
+
+    iterkeys = __iter__
+
+    def itervalues(self):
+        for key in self:
+            yield self[key]
+
+    def iteritems(self):
+        for key in self:
+            yield key, self[key]
+
+    def keys(self):
+        return list(self.iterkeys())
+
+    def values(self):
+        return list(self.itervalues())
+
+    def items(self):
+        return list(self.iteritems())
+
+    def __len__(self):
+        return len(self.keys())
+
+    def __contains__(self, key):
+        return any(key in d for d in self.dicts)
+
+    has_key = __contains__
+
     def __repr__(self):
-        content = dict.__repr__(self) if self else ''
+        content = repr(self.dicts) if self.dicts else ''
         return '%s(%s)' % (self.__class__.__name__, content)
+
+
+class CombinedDict(CombinedDictMixin, ImmutableDictMixin, dict):
+    """
+    An immutable :class:`dict` which combines the given `dicts` into one.
+
+    You can use this class to combine dicts of any type, however different
+    interfaces as provided by e.g. :class:`MultiDict` or :class:`Counter` are
+    not supported, the same goes for additional keyword arguments.
+
+    .. versionadded:: 0.2
+    """
 
 
 class MultiDictMixin(object):
@@ -307,9 +387,37 @@ class ImmutableMultiDictMixin(ImmutableDictMixin, MultiDictMixin):
 
 class ImmutableMultiDict(ImmutableMultiDictMixin, dict):
     """An immutable :class:`MultiDict`."""
-    def __repr__(self):
-        content = dict.__repr__(self) if self else ''
-        return '%s(%s)' % (self.__class__.__name__, content)
+
+
+class CombinedMultiDict(CombinedDictMixin, ImmutableMultiDictMixin, dict):
+    """
+    An :class:`ImmutableMultiDict` which combines the given `dicts` into one.
+
+    .. versionadded:: 0.2
+    """
+    def getlist(self, key):
+        return sum((d.getlist(key) for d in self.dicts), [])
+
+    def iterlists(self):
+        result = OrderedDict()
+        for d in self.dicts:
+            for key, values in d.iterlists():
+                result.setdefault(key, []).extend(values)
+        return result.iteritems()
+
+    def iterlistvalues(self):
+        for key in self:
+            yield self.getlist(key)
+
+    def iteritems(self, multi=False):
+        for key in self:
+            if multi:
+                yield key, self.getlist(key)
+            else:
+                yield key, self[key]
+
+    def items(self, multi=False):
+        return list(self.iteritems(multi))
 
 
 class _Link(object):
@@ -495,12 +603,26 @@ class OrderedDict(dict):
         return '%s(%s)' % (self.__class__.__name__, content)
 
 
+class ImmutableOrderedDict(ImmutableDictMixin, OrderedDict):
+    """
+    An immutable :class:`OrderedDict`.
+
+    .. versionadded:: 0.2
+    """
+
+    __repr__ = OrderedDict.__repr__
+
+
 class OrderedMultiDict(MultiDictMixin, OrderedDict):
     """An ordered :class:`MultiDict`."""
 
 
 class ImmutableOrderedMultiDict(ImmutableMultiDictMixin, OrderedDict):
     """An immutable :class:`OrderedMultiDict`."""
+
+    def __repr__(self):
+        content = repr(self.items()) if self else ''
+        return '%s(%s)' % (self.__class__.__name__, content)
 
 
 class Counter(dict):