+++ tests/datastructures.py
@@ -13,10 +13,21 @@ import random
 
 from attest import Tests, TestBase, test, Assert
 
-from brownie.datastructures import missing, MultiDict, OrderedDict, LazyList, \
-                                   Counter, ImmutableDict, ImmutableMultiDict, \
-                                   OrderedMultiDict, ImmutableOrderedMultiDict, \
-                                   OrderedSet
+from brownie.datastructures import (
+    missing,
+    LazyList,
+    OrderedSet,
+    MultiDict,
+    ImmutableDict,
+    ImmutableMultiDict,
+    OrderedDict,
+    ImmutableOrderedDict,
+    OrderedMultiDict,
+    ImmutableOrderedMultiDict,
+    Counter,
+    CombinedDict,
+    CombinedMultiDict
+)
 
 
 class TestMissing(TestBase):
@@ -72,6 +83,12 @@ class DictTestMixin(object):
         Assert(d[1]) == 2
         d[1] = 3
         Assert(d[1]) == 3
+
+    @test
+    def getitem(self):
+        d = self.dict_class([(1, 2), (3, 4)])
+        Assert(d[1]) == 2
+        Assert(d[3]) == 4
 
     @test
     def delitem(self):
@@ -225,6 +242,58 @@ class ImmutableDictTest(TestBase, ImmutableDictTestMixin):
     @test
     def type_checking(self):
         assert isinstance(self.dict_class(), dict)
+
+
+class CombinedDictTestMixin(object):
+    # .fromkeys() doesn't work here, so we don't need that test
+    test_custom_new = None
+
+    @test
+    def fromkeys(self):
+        with Assert.raises(TypeError):
+            self.dict_class.fromkeys(['foo', 'bar'])
+
+    @test
+    def init(self):
+        with Assert.raises(TypeError):
+            self.dict_class(foo='bar')
+        self.dict_class([{}, {}])
+
+    @test
+    def getitem(self):
+        d = self.dict_class([{1: 2, 3: 4}, {1: 4, 3: 2}])
+        Assert(d[1]) == 2
+        Assert(d[3]) == 4
+
+    @test
+    def get(self):
+        d = self.dict_class()
+        Assert(d.get(1)) == None
+        Assert(d.get(1, 2)) == 2
+
+        d = self.dict_class([{1: 2}, {1: 3}])
+        Assert(d.get(1)) == 2
+        Assert(d.get(1, 4)) == 2
+
+    @test
+    def item_accessor_equality(self):
+        d = self.dict_class([{1: 2}, {1: 3}, {2: 4}])
+        Assert(d.keys()) == [1, 2]
+        Assert(d.values()) == [2, 4]
+        Assert(d.items()) == [(1, 2), (2, 4)]
+        Assert(list(d)) == list(d.iterkeys()) == d.keys()
+        Assert(list(d.itervalues())) == d.values()
+        Assert(list(d.iteritems())) == d.items()
+
+    @test
+    def repr(self):
+        Assert(repr(self.dict_class())) == '%s()' % self.dict_class.__name__
+        d = self.dict_class([{}, {1: 2}])
+        Assert(repr(d)) == '%s([{}, {1: 2}])' % self.dict_class.__name__
+
+
+class TestCombinedDict(TestBase, CombinedDictTestMixin, ImmutableDictTestMixin):
+    dict_class = CombinedDict
 
 
 class MultiDictTestMixin(object):
@@ -394,6 +463,62 @@ class TestImmutableMultiDict(TestBase, ImmutableMultiDictTestMixin,
             assert isinstance(d, type), type
 
 
+class TestCombinedMultiDict(TestBase, CombinedDictTestMixin,
+                            ImmutableMultiDictTestMixin,
+                            ImmutableDictTestMixin):
+    dict_class = CombinedMultiDict
+
+    # we don't need this special kind of initalization
+    init_with_lists = None
+
+    @test
+    def getlist(self):
+        d = self.dict_class()
+        Assert(d.getlist(1)) == []
+        d = self.dict_class([MultiDict({1: 2}), MultiDict({1: 3})])
+        Assert(d.getlist(1)) == [2, 3]
+
+    @test
+    def lists(self):
+        d = self.dict_class([
+            MultiDict({'foo': ['bar', 'baz']}),
+            MultiDict({'foo': ['spam', 'eggs']})
+        ])
+        Assert(list(d.iterlists())) == d.lists()
+        Assert(d.lists()) == [('foo', ['bar', 'baz', 'spam', 'eggs'])]
+
+    @test
+    def listvalues(self):
+        d = self.dict_class([
+            MultiDict({'foo': ['bar', 'baz']}),
+            MultiDict({'foo': ['spam', 'eggs']})
+        ])
+        Assert(list(d.iterlistvalues())) == d.listvalues()
+        Assert(d.listvalues()) == [['bar', 'baz', 'spam', 'eggs']]
+
+    @test
+    def multi_items(self):
+        d = self.dict_class([
+            MultiDict({'foo': ['bar', 'baz']}),
+            MultiDict({'foo': ['spam', 'eggs']})
+        ])
+        Assert(list(d.iteritems(multi=True))) == d.items(multi=True)
+        Assert(d.items(multi=True)) == [
+            ('foo', ['bar', 'baz', 'spam', 'eggs'])
+        ]
+
+    @test
+    def item_accessor_equality(self):
+        CombinedDictTestMixin.item_accessor_equality(self)
+        d = self.dict_class([
+            MultiDict({'foo': ['bar', 'baz']}),
+            MultiDict({'foo': ['spam', 'eggs']})
+        ])
+        Assert(d.values()) == [d['foo']]
+        Assert(d.lists()) == [(key, d.getlist(key)) for key in d]
+        Assert(d.items()) == [(k, vs[0]) for k, vs in d.lists()]
+
+
 class OrderedDictTestMixin(object):
     dict_class = None
 
@@ -468,6 +593,20 @@ class TestOrderedDict(TestBase, OrderedDictTestMixin, DictTestMixin):
     def type_checking(self):
         d = self.dict_class()
         assert isinstance(d, dict)
+
+
+class TestImmutableOrderedDict(TestBase, OrderedDictTestMixin,
+                               ImmutableDictTestMixin):
+    dict_class = ImmutableOrderedDict
+
+    clear_does_not_keep_ordering = pop_does_not_keep_ordering = None
+    setitem_order = setdefault_order = update_order = None
+
+    @test
+    def popitem(self):
+        d = self.dict_class()
+        with Assert.raises(TypeError):
+            d.popitem()
 
 
 class TestOrderedMultiDict(TestBase, OrderedDictTestMixin, MultiDictTestMixin,
@@ -1058,5 +1197,6 @@ class TestOrderedSet(TestBase):
 datastructures_tests = Tests([
     TestMissing, ImmutableDictTest, TestMultiDict, TestOrderedDict,
     TestCounter, TestLazyList, TestImmutableMultiDict, TestOrderedMultiDict,
-    TestImmutableOrderedMultiDict, TestOrderedSet
+    TestImmutableOrderedMultiDict, TestOrderedSet, TestCombinedDict,
+    TestCombinedMultiDict, TestImmutableOrderedDict
 ])